+++ editor/compiler/interface/Runtime.cpp
@@ -15,19 +15,19 @@ void Runtime::runUpdate() {
     MaximFrontend::maxim_run_update(get());
 }
 
-void Runtime::setBpm(float bpm) {
+void Runtime::setBpm(double bpm) {
     MaximFrontend::maxim_set_bpm(get(), bpm);
 }
 
-float Runtime::getBpm() {
+double Runtime::getBpm() {
     return MaximFrontend::maxim_get_bpm(get());
 }
 
-void Runtime::setSampleRate(float sampleRate) {
+void Runtime::setSampleRate(double sampleRate) {
     MaximFrontend::maxim_set_sample_rate(get(), sampleRate);
 }
 
-float Runtime::getSampleRate() {
+double Runtime::getSampleRate() {
     return MaximFrontend::maxim_get_sample_rate(get());
 }
